refactor: LottoStatistics에서 상금을 Money 객체로 생성하고 관리하도록 변경

+++ src/main/java/lotto/domain/Rank.java
@@ -31,7 +31,7 @@ public enum Rank {
         return getRank(matchCount);
     }
 
-    public double calculateTotalReward(Integer count) {
+    public int calculateTotalReward(Integer count) {
         return reward * count;
     }
